[fuchsia-async] Fix hang in TCP connect
When connect returns EINPROGRESS, clients are expected to wait for the
socket to become writable and then query the socket's error state to
determine if the connection attempt succeeded or failed. This change
modifies TcpStream to check the error state after the socket becomes
writable, and adds a test to verify connection attempts to nonexistent
endpoints do not hang.

Bug: PKG-873 #done
Change-Id: Iba57c0cd5f3f80e65bf20aa02fd77b48b7539951