[appmgr][svchost] Route fuchsia.boot.RootJob to the sys realm.
This routes these services to the sys realm and migrates users in
fuchsia.git to use them instead of /dev/misc/sysinfo.

In an effort to have better process isolation, we are trying to remove
GetRootJob and GetRootResource from /dev/misc/sysinfo. Doing that today
is infeasable due to a number of reasons. Thus, in the mean time we are
going to make this more auditable by using more appropriately named,
fine-grained services that vend out these capabilities directly and put
them behind an allow-list.  Note that these allowlists do not capture
things in zircon/system/ since they don't use component manifests.
Our ultimate goal will be though to remove as many usages of these
services as possible or to compile them out of the system when building
something that will be shipped to actual users.

SEC-323

Change-Id: I24ba44198e5c49462115e82a4c0c46dfea759a25