[zircon][fidl] Add metric broker interface.
This interface provides an opt-in mechanism for instrumented processes
to delegate the responsibility of logging remote metrics to another
service.

The goal is to reuse the inspect-API vmo and use user written
configuration file to map a metric path to a cobalt metric. This will
reduce the boilerplate developers need to deal with to instrument both
local an remote metrics. Plus reduces the amount of IPCs instrumented
processes need to perform.

TEST=builds

Change-Id: Iece906e02af37ca8e6a2225822e393aa07c57ca3