[debugger] Print local vars with inlines.
Fixes a bug where the current frame is a physical frame not at the top
of the stack, but the current code location is in an inlined function.
Previously GetMostSpecificChild() would actually go into the inlined
function so searching for the variables in the current scope actually
just searches the inlined function.

Adds a parameter to control whether GetMostSpecificChild() goes into
inlined functions. Although we never set this to true currently, I made
it optional because it seems conceptually weird for a function with this
name to not be able to go into inlines (uses that don't go through the
elaborate Stack object will want this behavior).

Change-Id: Iff92ca9322d8176054cc9893be6cea6cd0eba43d