[kernel] Manipulate WaitStateObserver::wakeup_reasons_ with atomics
The WaitStateObserver's wakeup_reasons_ member type is mutated from
multiple threads through the StateObserver callbacks. The mutating
callbacks are serialized through one dispatcher lock, so only one will
occur at a time.  When one of these mutations trigger a wakeup, an event
is signaled which will lead to the End() call executing on a different
thread from mutators.  It's important that the End() call observe a
value that was set and that was sufficient to trigger a wakeup.  The
specific value observed by End() doesn't matter in the case of repeated
mutations.

This upgrades the wakeup_reasons_ type from a raw primitive
(zx_signals_t) to an ktl::atomic<> and replaces the stores and loads
with atomic operations using the acquire / release memory order.

This is a reland of 73f6bfb609d5b9a56dcf1f1681bd204d75dcc64c, but
without the debug assertion in object_wait. That assertion seems to be
triggering for unrelated reasons and is being investigated in ZX-4697.

ZX-4670 #done

Change-Id: Ib406ab7824e6133b9a38024023dd3e21e2d2a87b