[debugger] Use expressions for more formatting.
Everything but C++ std::string can be expressed generally as a pointer
and a size. Replace those implementations with new ones that use string
expressions for pretty-printing the types.

Use PrettyArray to pretty-print C++ std::vector.

Make the PrettyTypeManager more general using a string prefix lookup.

Add a pretty-printer for a Rust BinaryHeap.

Change-Id: Icf434356895e5acc522cba489b17c277bb8776f9