[wlan][iwlwifi] Port pcie/rx.c
Convert spinlock functions to mutex.
Convert list functions.
Convert Linux DMA api to io_buffer.

The Linux driver separates the DMA buffer allocations from mapping it to
the device. This is so that when a page is "stolen" it can simply unmap
the page and pass it up the stack. In Fuchsia, it's simpler and
potentially faster to simply copy data out of the VMO instead - although
we should investigate the performance further. For now, we do not
separate creating the VMO from pinning it.

TEST=fx run-test iwlwifi-test. Added unit tests that check that the
initial state is correct.
BUG=WLAN-895

Change-Id: I8692051780eff0ddbf268a47588457d73617d49e