[kernel][vm] Convert recursive function to iterative
When a COW clone is closed, it's necessary to release pages in ancestor
hidden vmos which are no longer referred to by any non-hidden vmo. The
original implementation did this by recursively walking the clone tree.
This change replaces the recursive walk with an iterative walk. To
facilitate backtracking during the walk, it is necessary to store a
small amount of state (64 bits). This state is stashed in the VmObject
itself instead of on the stack, to preclude blowing the stack.

ZX-1268 #comment [kernel][vm] Convert recursive function to iterative

Change-Id: I5811d1368af372784b76e08dc9dd0d8d37df7360