[fidl] add generation for encoder/decoder (c++).
Currently, the only way to encode a request or a response is to use the
sender classes.

This CL adds stand alone encoder functions for requests and responses.

This is needed if a program (or a library) wants to a fine control on
messages. For example if the programs wants to use zx_channel_write_etc.

Factorize a little bit encoding and decoding. The size for echo programs
in C++ is reduced by 0.23% (and we have now the stand alone encode decode).

Change-Id: Ic070e552e638e0c8622ddcf22a50783df6b90ea7