[kernel] Use GPArena for handle table arena
Changes from fbl::Arena to fbl::GPArena for the handle table arena.
The GPArena allows for explicitly guaranteeing that parts of the Handle
will be preserved whilst it is unallocated and has internal
synchronization, allowing for removal of the ArenaLock.

Full performance data is in ZX-3426. The summary of the results is that
for operations that are short and create/lookup handles then
performance improvements are 5-15% decrease in runtime. These tests are
things like: handle creation, deletion, object wait async, small
fifo/socket/channel reads/writes.

ZX-3426 #done
ZX-95 #done

Change-Id: Idc67deefb717639cd96ce75e74d9e2a474df952f