[core-tests] Fix vmo kmem tests not running
Although the tests compile and run if a using directive is used to
rename a test fixture, the setup method is not run, which causes the
tests to be skipped.  This change simply renames the fixture to the
desired name.

This test also runs the utests before the zxtests in core-tests, as
cleanup  of the utest infrastructure can interfere with the kmem
accounting in the vmo tests.

Change-Id: I5d9366bcf1d2f6b3422ad82dabb99f375a280274