[workstation] Add audio/audio_core packages to cache.
Currently the audio/audio_core packages are only listed in
"universe_package_labels", meaning the audio is not available on boot
until the packages are fetched from the network. Move them to
"cache_package_labels", so that we can immediately play audio after
boot.

We may want to move it to base longer term, but we can hold off that
decision for now.

Change-Id: I60951fb4c6795a0f1c32afbaa979ec63ed8fa8b8