[debugger] Correctly check when downloading binaries is needed
We always get the symbols first now, and then we check to see if the
binary is needed as well. This fixes a bug where we'd ignore the
binaries entirely if the symbol server came online after the module was
loaded, and it's less noisy on symbol servers that don't provide the
binaries but send unstripped binaries, so we don't need them.

Change-Id: Ieb5d36e7d5e32e40cd6798b8620aead8c726070c