[ledger] Allow ObjectIdentifiers to outlive their factory
This turns out to be necessary because ObjectIdentifiers are captured in
tasks that are posted to the dispatcher, that may outlive even
LedgerRepository.

Add a factory() method to ObjectIdentifier to allow checks of whether
an object is used on a storage with a different factory from the one that
emitted it (or whether it is still tracked).

Bug:  LE-702
Test: fx run-test ledger_tests -t ledger_unittests
Change-Id: I9649ad3687879694436f8e1b5612d0bbcb91ca0d