[ledger] Use ObjectIdentifierFactory in encryption service
As a result of using the factory in the fake encryption service, a
lot of fakes, test utility functions and tests need to be updated to
use an object identifier factory as well because they rely on the fake
encryption service to allocate fake object identifiers. Note that fake
object identifiers may be tracked by a real factory in some tests,
so it is not an option to make all fake object identifiers untracked.

Bug: LE-702
Test: fx run-test ledger_tests -t ledger_unittests
Change-Id: I36b9d4809379db03debfc0f79d59b4ef95cf613a