[fidlgen][fidlc] Correct identifier references in JSON IR
Identifiers referencing values in external libraries are incorrectly
given a name in the current library. This corrects the naming and
updates the C++ generator to use the member field.

In https://fuchsia-review.googlesource.com/c/topaz/+/301408, dart
was changed to support generation once member name was supplied.
The dart behavior will change once this CL is in and the member is
populated.

[FIDL-735] #done

Test: fidlc tests, dart fidl tests, C++ fidlgen tests

Change-Id: I2dcd8acb6b10c87347e3aca9050ddf289221b0f8