[docs] Add an overview to the Fuchsia Archive format docs
A Fuchsia Archive is just a way to represent a mapping from strings
(pathnames) to strings (file contents).  It's useful to state that
explicitly so that one does not need to read the whole doc to be sure
that that is true.

Also mention "FAR" and ".far" -- they were not otherwise mentioned in
this doc, but this is how the format is referred to elsewhere.

Issue: none
Test: none
Change-Id: Ib2480c141dbbada36dbd95c474fa288c29310943