[virtualization][tests] Add TerminaEnclosedGuest.
Add a new guest system that uses the termina_guest to run tests. This
will enable integration tests of both virtio_wl and virtio_magma
devices.

One primary deviation from the Zircon and Debian guests is that the
Termina guest will rely on vsock services to communicate with the guest
instead of a serial console.

This change is primarily focused on the test infrastructure for the
termina_guest. Adding more tests that target this system will come in
follow up changes.

MAC-218 #comment

Test: fx run-tests guest_integration_tests
Change-Id: I0323648551e1c8fc3ffa2e2ae41b749531d196f6