[magma] add icd manifest generation script
This change adds a script to generate a vulkan ICD manifest using the
given path string. It also updates the magma build files to use the
script instead of string literals, and adds a manifest file to the linux
guest extras partition.

The change also adds a new build-time test that compares the output of
the script to a reference manifest file.

Test: ran new icd_manifest_compare build-time test
      ran the script with various arguments and inspected output
      ran virtmagma_vulkan_unit_tests, which uses the manifest file
Change-Id: I5ea6838b6e0c39bc02857271a50142916402a620