[mediacodec] Implement AAC Encoder
This CL implements an AAC encoder codec adapter using FDK.

This encoder uses a dispatcher for processing input and an atomic flag
for stopping in order to prevent input from processing after a stop
call. This is more responsive than the old method which could
potentially spend time decoding before terminating on requesting an
output buffer.

TEST: fx run-test audio_encoder_test
Change-Id: Ib3f286cd810a4d22184c314647b664d7a7d67550