[mdns] filter out NIC-to-NIC loopback messages
The mdns implementation handles an arbitrary number of NICs. When two
or more NICs are connected to the same network segment, however,
messages sent out one of the NICs and arriving on another NIC weren't
being filtered out properly. This has been fixed in the message
transceiver.

In order to test for regressions, mdns_netemul_tests was modified to
have two NICs on the 'local' end of the conversation. This repros the
bug effectively without the fix and verifies that the fix works.

Having two NICs on the local side also revealed an issue with the
host name resolver that occurred when host name queries were answered
on the two NICs concurrently. This has been fixed as well, and the
updated mdns_netemul_tests tests for regressions here as well.

NET-2458 #done

Change-Id: I11c06a03b9abedd20ca260986fee6ed7d14a2da7