[perfcompare] Print a second table, listing only test cases with differences
Before this change, perfcompare prints a single table listing all test
results, and you have to search through the table (e.g. with Ctrl-F in
the browser) to find improvements and regressions.

After, perfcompare prints two tables: The first lists only tests that
changed, and the second (as before) lists all the tests.

This should make perfcompare's output more usable!

Change FormatTable() to take the heading row as a separate argument
rather than as the first row in the rows list.

Issue: none
Test: "python perfcompare_test.py" + optional perfcompare trybot
Change-Id: I4ca705694305f753ea93118fdfd84671cbb66254