[tools/blackout] add blackout library for power failure testing
this directory contains the blackout library, which provides
functionality for both the host-side and target-side of filesystem power
failure host-target interaction tests. it also contains the first test,
which is a test for blobfs with a soft reboot using fsck for
verification. this test uses a load generator utility (written in C++ at
the moment) for creating, writing, reading, and deleting random files on
a provided blobfs partition. this load generator utility is run by a
rust shim that manages the blobfs partition.

it is not included in the base build by default. to build it,
//tools/blackout:all has to be added as an additional package to fx set
in the usual way.

temporary details:

the test expects an additional, empty blobfs partition to be present. it
is not present in a normal build. the best way I've figured out to get
it there consistently is a modification of build/images/BUILD.gn to
create an empty partition for me, but that's not a good general
solution (and is not included in this changelist). I will be
investigating using netbooting to get around this particular constraint.

ZX-4289 #done
ZX-4290 #done
ZX-4292 #comment in progress
ZX-4293 #comment in progress

Test: blobfs-fsck-soft-target-test (for load generator)
Change-Id: I1d4d6cd8b1bb4854317935f7a0ca06729c13188c