[debugger] Generalize array printing.
Adds support for evaluating expressions in pretty-printers. Adds a
general "array" pretty-printer that takes expressions describing the
array, and uses these to format the Rust vector.

In the future, we can have a more extensible way for describing these
expressions but I expect we can reuse the same machinery.

It is currently not possible to descibe C++ vectors yet because it
requires pointer subtraction which is not yet supported.

In support of this, adds a way to evaluate multiple expressions and call
a callback when they're all done.

There should be no observable behavior change.

Change-Id: I98b065021b81e6293e5c1b5d3fe8524f7dbd32ef