[optee] OpenFile should return TEE_ERROR_ITEM_NOT_FOUND
Trying to open a file, or creating it upon failure to open is a common
idiom that should be supported. Throwing `TEE_ERROR_BAD_STATE` is
nonideal in this situation, since the TA may consider it an
unrecoverable error (especially since it does not exist as a possible
return code in the contract of TEE_OpenPersistentObject).

Returning `TEE_ERROR_ITEM_NOT_FOUND` in such a situation is more
consistent with the TEE Internal Core API Specification.

Test: Added test case for desired behaviour in optee_test

Change-Id: Ie13a6279d64e20c1d63cd3fcc73feb438dc482ed