[devshell][test] bash_test_framework with mock features
Adds a test framework for bash scripts.

As discussed, this CL starts with a patchset of the original snapshot
from fxr/281691, as of the last patchset pushed (June 12, 6:15pm).
The next patchset shows the changes and additions.

Test functions run in isolation, with a specific, limited set of files
and executables from a source project. Executable binaries or scripts
can also be mocked rather than copied, to reduce dependencies on complex
programs that should be tested independently, reduce resource
requirements, and/or avoid building tools that are not strictly required
for testing the target script.

Also squashed a follow-up CL into this one (as discussed with raggi@):

* Abort test after attempting first test function, if there is an error
  during initialization or set up. Return status 255 (MAX_ERROR_STATUS)
  now indicates a set up error, as opposed to a test failure count.
* Framework error messages display the source file and line of the
  test code that precipitated the error, making it much easier to find
  the root cause in most cases.
* Added a post-set-up step to validate mock executable extension file
  names. While writing tests, this was a common issue (even for the
  framework author) so bad extension names are identified and fail fast.

Test: fx run-bash-test bash_test_framework_test

Change-Id: Ic85043ed66561563a09a05b41024247ff14359d4