[component] correctly populate dirent.Type
This changes the Go "pseudo directory" fuchsia.io.Directory.ReadDirents
implementation to use the correct type enumerations. Before this change,
the implementation would always set the type to zero because the values
used had none of the low 8 bits set.

A test is added to exercise this code from Rust, where the directory
traversal implementation relies on the type enumeration (as opposed to
relying on fuchsia.io.Node.GetAttr, as some other implementations do).

NET-2516 #done

Change-Id: Idc6c3531f7dddd86e9eb0c434d28c7cac16eadde