[mdns] Implement IPv6 listening.
Modify the implementation to support listening on IPv6
multicast addresses as well.

Add another question record to each QuestionPacket()
result to ensure each one asks for both the IPv4 and IPv6
address of a given domain.

Note however that AnswerPacket() creates a packet with a
single answer record, whose record type (A or AAAA) depends
on the configured IP address.

+ Modify dev_finder to listen on both IPv4 and IPv6 addresses.

+ Modify mdnstool so that the "publish" and "resolve" command
  support both IPv4 and IPv6 addresses. Note that "publish"
  can support only one address type at a time.

BUG=DX-1664

Change-Id: Iddc9439f86c1958a131bf07e0b856e8bdcb2f85d