[debugger] Use globs for pretty-printer matching.
Previously the code used prefix matching to trigger specific
pretty-printers but this has significant limitations. This patch instead
uses the new TypeGlob for matching template parameters in a type-aware
manner.

Adds the ability for a TypeGlob to be constructed with a manually
created ParsedIdentifier for things that can't be expressed using the
normal identifier parser.

Adds a test for std::vector<bool> that the types are matched in the
correct order, even though vector<bool> pretty-printing isn't
implemented.

Change-Id: I6a713a25cf9398c272999d3e0027762f234bd114