[debugger] Better detect unstripped binaries and use them where we can
The symbol server, and the build ID repo, offer locations for a debug
symbol file, and for a file containing the original binary. However, an
unstripped binary can appear in either location and potentially provide
both.

This patch shifts to a paranoid policy. We almost never trust contextual
information on what a file contains. If it passes into zxdb, we open it
with elflib and probe it to see what it offers. Furthermore, we always
try to download the symbols first, and if probing them reveals they're
unstripped binaries, we avoid the binary download, which spares us a
whole pile of spurious error messages.

Change-Id: Icdece7e2db56c8bbe0afc039c031eccc47cd5b33