[perfcompare] Run perfcompare_test.py on the bots as a host side test
This involves adding a wrapper program, written in Go, for invoking
Python.

This is based on a similar Go wrapper for running Python tests in
src/lib/fuzzing/host/host_test.go, which was added in this change:
https://fuchsia-review.googlesource.com/c/fuchsia/+/276943

Before, perfcompare_test.py did not get run on the bots.

Issue: PT-203
Test: local and fast:
    fx set bringup.x64 --with //garnet/bin/perfcompare:tests
    fx run-host-tests host_perfcompare_test
  local to check what bots do (slower):
    fx set bringup.x64 --with //bundles/buildbot:core
    fx run-host-tests host_perfcompare_test
  via trybots:
    look at fuchsia-x64-debug trybot results
    & check that "host_perfcompare_test" appears there
Change-Id: I7c149edb1f5e5732d54c1be0c3684f74f70169a9