[netstack3] Add utilities for execution context
This commit introduces the `context` module. It contains traits
for various kinds of "execution contexts". Eventually, the plan is
to replace the current `EventDispatcher` family of traits and the
`Context` type with these traits.

The primary benefit of this change is that it will allow us to
use what we refer to today as the "core/bindings distinction"
pattern throughout the stack. Currently, we have only one such
distinction - between the actual bindings (in netstack3/src) and
the core (in netstack3/core/src). Eventually, we plan to model many
pieces of the stack as their own, smaller, "cores". This affords a
number of advantages:
- Tests become simpler because we don't need to mock unnecessary
  parts of the stack. For example, when testing the ARP module,
  there need only be a single copy of ARP state. We don't need any
  devices, IP forwarding tables, transport layer sockets, etc.
- We get better code reuse through factoring APIs into traits that
  capture common patterns, such as the `StateContext` trait for
  accessing state or the `FrameContext` trait for sending frames.
- We are forced to reason more precisely about our API boundaries.
- We are better prepared for being able to produce binaries that
  only make use of a subset of the available functionality - for
  example, binaries which are IPv4-only, or which do not support
  TCP, etc.

Test: No tests so far, as there are only trait definitions, and
      no executable code.
Change-Id: I142635f7d5738a4e0137752aacd1a6c1fa33109b