[kernel] Avoid allocation in pmm callback
The callback given to |pmm_init_reclamation| is called under the pmm
lock and needs to not cause memory allocations or reclamations itself.
|user_signal_self| may need to perform allocations, however, and this
could lead to calling back into the pmm. This was detected by lockdep
with the following trace:

[00006.522] 00000:00000> ZIRCON KERNEL OOPS
[00006.522] 00000:00000> Circular lock dependency detected:
[00006.522] 00000:00000>   lockdep::LockClass<TheHeapLock, Mutex, 0, lockdep::LockFlagsSingletonLock>
[00006.522] 00000:00000>   lockdep::LockClass<PmmNode, Mutex, 102, lockdep::LockFlagsNone>
[00006.522] 00000:00000>   lockdep::LockClass<SoloDispatcher<PortDispatcher, 32783, 0>, Mutex, 272, lockdep::LockFlagsNone>
[00006.522] 00000:00000>   lockdep::LockClass<SoloDispatcher<EventDispatcher, 53251, 8>, Mutex, 272, lockdep::LockFlagsNone>

This change preserves the existing semantics and makes the callback
yield the memory state index to the oom thread, which now signals
the user as well as performing reclamation.

Test: Previous functionality preserved, lockdep violation no longer reported

Change-Id: I4d7fcc877326044e73efcdcea508a4317e35b8a2