[build][gigaboot][kernel] Enable profile build of Gigaboot
The profiling/coverage instrumentation works for the Windows/EFI
target too.  Adjust the "runtime" (really link-time) data header
in the kernel implementation for the EFI build environment and
use it in Gigaboot as well.  Later work will actually transfer
the collected data on where it can be used.

This also disables the "profile" variant for the x86 kernel.  The
link is broken due to TC-567 (compiler instrumentation issue).

Bug: TC-598 #comment enable profile build of Gigaboot
Bug: TC-567 #comment disable profile build for x86 kernel
Test: fx multi set bringup.{arm64,x64} --variant profile && fx multi build
Change-Id: I6e00e74d035d567f6fa9ee7c13ab84cffd381f35