[build] Add zbi_compression build argument
This selects between lz4f and zstd for ZBI item compression.
This affects what the build creates and what userboot can decode.
It can also choose the compression level.

This also renames "lz4" to "lz4f" in the zbi tool's code and --compress
switch argument syntax.  These are two different formats and lz4f is the
proper name for the one we actually use in the ZBI format.

Bug: ZX-4769 #comment zbi_compression build argument
Change-Id: I72284e7f08d6a3313e65373ccdd864ae3a979712