[debuglog] Shutdown debuglog before mexec.
Debug log was not shutdown before an mexec and this seemed to be causing
a panic when the new kernel starts. Shutting it down explicitly seems to work around the issue.
Invesitgation is still ongoing as to why we can't just depend on the new kernel coming up to
reset all state.

Test: Netbooted sherlock 10+ times with no panic on boot.
Issue: ZX-4687
Change-Id: Idb11c15aa01b1377ac21623fc23d1bbdab91b78b