[fidl][hlcpp] Begin fidl::*Ptr to fit::optional<std::*> move
This makes StringPtr and VectorPtr offer similar APIs to
fit::optional<std::string> and fit::optional<std::vector> and optionally
marks parts of the StringPtr and VectorPtr API that won't be available
as deprecated.

The biggest thing that isn't caught is when a StringPtr is explicitly
converted to a std::string in a system header the deprecation warning
isn't shown. This happens in std::make_unique<>() and
std::map<>.emplace().

FIDL-564 #comment

Change-Id: I9c09fec1295a8a7ef463dffdb542685066e7ce11