[media][tests] Run loop exit logic on the async thread.
It was possible that the thread running the loop would exit after the
loop was moved into the shutdow state (which would in turn run the
async::Loop dtor and close the port handle) and before the loop could
post a wake task, which would return ZX_ERR_BAD_STATE from
zx_port_queue and trigger an assert.

By posting the quit task to the loop, we ensure that this will not race
with the loop exit logic.

Test: Was causing flakes in fxr/300198. This will be added to the
      relation chain add will do some burn-in testing in CQ with
      that CL.
Change-Id: Iad5c63456717346a028c4600ed2cacba90d63a0b