[perfcompare] Require the multi-boot perf results format as input
This addresses the TODO in ResultsFromDir().  The
fuchsia_perfcompare.py recipe is now generating results for multiple
boots of Fuchsia, using the new format (format #2), so we can drop
support for reading results in the old format (format #1).

This requires changing the test code a bit, because it blurred the
lines between the old and new formats:

 * Change WriteExampleDataDir() to generate the new format.

 * Remove test_reading_results_from_multi_boot_dir() -- it was only
   applicable when ResultsFromDir() supported reading two formats.

 * Rewrite test_reading_results_from_tar_file() as
   test_readback_of_data_from_tar_file().

 * Update test_validate_perfcompare() because validate_perfcompare
   still takes the per-boot directories (format #1) as input.

Issue: PT-202
Test: "python perfcompare_test.py" (run by bots on CQ)
Change-Id: Ic383dd5dc43e9c4fb7834c3ad86cf67a45d70cc2