[perfcompare] Switch to better handling of cross-boot variation
Change the t-test CI calculation so that the sample size is the number
of boots per revision, where each value in those samples is a mean of
all the runs within a boot.  This should generate wide-enough CIs and
hence avoid unnecessary Type I errors for tests with non-trivial
amounts of cross-boot variation.

This addresses the TODO in ResultsFromDirs().  The previous behaviour
was a temporary hack which we can avoid now that the
fuchsia_perfcompare.py recipe and the test code have been updated to
produce multi-boot results.

Update comments in perfcompare.py.

I tested this by running the tool manually on a corpus of perf results
data from one revision of Fuchsia.  This change decreases the mismatch
rate (the rate of Type I errors), at the expense of increasing the
relative CI widths (increasing the rate of Type II errors):

Before:

  $ python perfcompare.py validate_perfcompare --group_size=5 out_perf_results_gz/*/0/out.tar.gz
  ...
  Mean mismatch rate: 0.071313
  Mean relative confidence interval width: 0.037157
  Number of test cases: 617
  Number of result sets: 4 groups of 5 boots each (ignoring 0 leftover boots)
  Expected number of test cases with mismatches: 44.000000

After:

  $ python perfcompare.py validate_perfcompare --group_size=5 out_perf_results_gz/*/0/out.tar.gz
  ...
  Mean mismatch rate: 0.000000
  Mean relative confidence interval width: 0.091805
  Number of test cases: 617
  Number of result sets: 4 groups of 5 boots each (ignoring 0 leftover boots)
  Expected number of test cases with mismatches: 0.000000

Test: "python perfcompare_test.py" (run by bots on CQ)
  + manual test above
Change-Id: I5310e84b954ad0047aa4703434ebc803a4217172