[componentmgr] RELAND "Implement framework capability to propagate hub listing"
This relands commit ff0751ac3855ad390b4bdc799ac7434469dfbacb

This CL implementing a HubReport framework capability via a hook to
allow hub_client to pass its view of the hub back to the integration
test for verification. This is an example of using hooks to implement
test-only framework capabilities.

This reland fixes a deadlock where HubTestHook was holding on to the observers
lock while it was waiting for a message from hub_client. However,
sending a message from hub_client needs to grab the observers lock as
well causing a deadlock. The solution is to simply transfer ownership of
the receiver from the shared Hashmap to local function until a message
is received and then transfer ownership back into the Hashmap.

Bug: CF-541
Change-Id: I4c62d13b5f669f34541b07a0f4ee1708b54e542d