[debugger] Add a globber for template types.
This will be used for the pretty-printing system to match type names.
Prefix searching and standard string-based globbing won't work for this
use-case as described in the new type_glob.h.

This adds a type matching and scoring system specifically for template
parameters. It is not currently hooked up to anything.

Change-Id: Ib9cedd473142abc1f743ccc261e2d95ccc53df66