[inspector] Don't output old backtrace format
This has been on the todo list for a while. The old backtrace system
doesn't even work and we've been getting along without it. I'm not
yet removing the contextual information because we still need it for
debugging. The old backtrace information contained stack pointers
which the symbolizer has supported for a long time but we never used
it. This change adds the stack pointer to the markup and removes the
old backtrace information.

Change-Id: Id4d4d6e3e1b8bea124f76ad977e2d9a2511fa3a9