[netstack3] Core allocates socket IDs, not bindings
Previously, the bindings allocated socket IDs and passed them to
the core when creating new sockets. This required a) the bindings
to manage its own IDs and, b) the core to accept any ID type, which
in turn required that sockets be stored in a HashMap.

With this change, the core stores sockets in an IdMap, and is
responsible for allocating IDs. This makes the bindings simpler
since they don't have to worry about allocating IDs, and makes
lookups faster since IdMaps are just vectors whose IDs are simple
indexes.

Test: Core's ICMP socket tests and Bindings' ping tests exercise
      this functionality.
Change-Id: I119d6285a5f793645676fbc3119b5545b2f08b4a