[debugger] Fill out accessors.
Fills out all remaining PrettyType accessors for the existing supported
types. The "empty" functions are commented out because our expression
language doesn't support "==" yet. This will be added at some point in
the future.

Adding these to the std::string PrettyType makes it even more complex
and unfortunate. I believe this is the last bit of complexity that will
beed to be added and hopefully we can simplify this soon.

Change-Id: Ia7e275007d31b873b477ab874f5579bafe47a655