[ledger] Remove unnecessary set_on_empty calls
Setting the PageConnectionNotifier's on-empty callback better fits the
"set the on-empty callbacks immediately after construction" practice
that is the norm throughout the rest of Ledger.

Setting the ActivePageManager's on-empty callback is already done when
the ActivePageManager is shortly after its own construction passed to
the ActivePageManagerContainer; there's no need to re-set it to an
equivalent value if the ActivePageManagerContainer's set_on_empty
method is called later.

Change-Id: I0afb81518a1d1e80d47df2a416db5b2da253d197