[debugger] Improve settings, add build-dirs
The console set/get command didn't work properly in the presence of
hierarchical settings that existed on multiple levels.

Fixes the issue and improves the code a bit. Improves the output by
adding some colors and indenting.

Adds a build-dirs setting. This is not currently hooked up. It will be
hooked up in a later pass but adding it now makes the test better.

Removes the --system switch to get/set because now there is a "global"
noun.

Add "adapters" from Chromium which provides a reverse-iterator adapter.

Change-Id: Idb2290e32b42fa7fa4049bf3357538624e5593e2