[inspect][kcounter] Expose kcounter to inspect
Adds fuchsia.kcounter FIDL as a miscsvc-hosted service which is a
distinguished userspace program that's allowed to directly read the VMO
format that the kernel exposes for kcounters. It does some summarization
and filtering to things we believe we might care about, and in turn
exposes the resulting data as an Inspect-formatted VMO.

Additionally, add a new component "kcounter_inspect" which connects to
this endpoint, requests the data and exposes it for retrieval by iquery.
This setup is required because only components may publish to /hub
currently.

A new test package is added as kcounter_inspect_tests, which currently
contains integration tests confirming that the FIDL endpoint can be
connected to, and that some standard kcounter data can be retrieved from
it, as well as a unittest binary of the binding helper.

Example, in one shell run:

  $ run kcounter_inspect.cmx

then, iquery to find the node, which will be something like:

  $ iquery --find /hub
  ...
  /hub/r/sys/6340/c/kcounter_inspect.cmx/12950/out/objects/root.inspect
  ...

and finally:

  $ iquery --recursive /hub/c/kcounter_inspect.cmx/12950/out/objects/root.inspect
  kcounter:
    handles:
      handles.duped = 4246
      handles.live = 3857
      handles.made = 20533
    channel:
      channel.messages = 28480
  ...

ZX-3324 #comment [inspect][kcounter] Expose kcounter to inspect

Change-Id: I1311c96b7c456beb63e99cc7fe70e1d2c4b9b5c7