[zircon] Separate out mapping logic in GetInfoForUserspace.
In particular, split out the functions that map internal kernel values
to external userspace constants from the GetInfoForUserspace() function.

The existing function also attempts to perform the mapping outside of
the thread lock. Given the mapping is a relatively trivial operation,
this CL simplifies the code to write results directly out to the "info"
struct.

Change-Id: I2a66db47b0a417d9a6e4ba00b786f4ede19230b4