[audio_core] Fix overactive DCHECK
The fix for MTWN-395 is simply to adjust an overly aggressive DCHECK.
Although it is reasonable to require mix jobs that call the
resamplers to limit their frac_src_position parameter to the
range used to determine when the previous mix job ended, it actually
is possible for a packet to be submitted that exceeds that position,
as long as the position is still less than the total packet size.
Code inspection confirms that if frac_src_pos is in this additional
range, the function runs (doing very little, naturally) and exits
without error. Both of the existing resamplers have been updated.

As an additional defense in depth measure (and actually an
optimization, since calling the mixer with this packet and timestamp
is unnecessary), the logic in audio_output has been reworked to
automatically skip the packet in cases like this.

Test: New test cases added

MTWN-395 #done

Change-Id: I6b33e81da1c611f084f06fb705bff8085b8971a2