[utest] Add more detail about core test failures.
We currently run two test frameworks for the core test: zxtest, followed
by unittest. If zxtest has a test failure, but unittest does not, the
tail of the log message is as follows:

    ====================================================
    Results:
        SUCCESS!  All test cases passed!
        CASES:  21     SUCCESS:  21     FAILED:  0
    ====================================================

which is understandably confusing to readers attempting to determine why
integration tests have failed.

This CL adds some more output, adding start/end messages before each
test framework runs, and also adding a failure summary if either of the
test frameworks fail, so the final messages in the log are "FAILED", and
not unittest's "SUCCESS" message.

Change-Id: I57b01ba8f0aec07a0f0fa0deb1424f9fef152ef7