[escher] Make Escher Geometry Its Own Lib
Escher geometry classes are now their own separate
library that can be built on their own without relying
on the rest of Escher nor Vulkan.

This is part one of an ongoing effort to decouple escher
and vulkan from the Scenic scene graph.

Testing - This is a pure refactor that does not add or
remove any functionality.

SCN-1531 #done
SCN-1530 #comment

Change-Id: I98d7cf705c60c5424d89821a5a164fcefa158bf4