[system-update-checker] update test for run_until_stalled changes
poller::test::test_simultaneous_user_update_check_and_periodic_update_check:
This test interleaved two executions of try_start_update by assuming
that the future given to run_until_stalled would be polled before any
message on the executor's port was processed (including from previous
fasync::spawn). This updates the test to not rely on this to create the
desired interleaving: instead, the update is explicitly blocked by
taking a lock, and is unblocked once the lock is released.

Change-Id: Ie7469e55a677aa585a3b6ecd9c4e2f9522393bd5