[kernel] Change pmm command for triggering oom
The existing k command to change the mem_avail_state is not very useful
for triggering oom scenarios since the pmm may notice there are lots of
free pages and immediately leave the oom state.

Instead, this adds a command that actually allocates and leaks pages to
cause a real oom scenario.

FLK-476 #comment

Change-Id: I6ebc8f8f66de7248396f0af15253c28c3d9f2391